re designed logger and fixed buffer clear
