+++ gatality.log
@@ -1,0 +1,5 @@
+[Info] wooho
+[Info] woohotwo
+[ERROR] oopsie
+[FATAL] big oopsie
+[Warning] no oopsie?
