+++ src/logging/logging.h
@@ -1,0 +1,9 @@
+#ifndef logging_h
+#define logging_h
+
+void logInfo(const std::string& message);
+void logWarning(const std::string& message);
+void logError(const std::string& message);
+void logFatalError(const std::string& message);
+
+#endif /* logging_h */
